Breakfast: Bake: Ignore smd_rpc_sym for Flo
Nexus 7 2013 (Flo) has this file in its kernel source. It's not a big
problem, though it's unlikely to be deleted, because this will make
setlocalversion set -dirty flag to a built kernel.

+++ core/source.sh
@@ -197,7 +197,7 @@ clean() {
 		clean_junk
 		hard_clean
 
-		if [ "$ARCH" == "arm" ]; then
+		if [ "$ARCH" == "arm" ] && [ ! $CUR_CONFIG == "flo.conf" ]; then
 			rm -f $KERNSOURCE/arch/arm/mach-msm/smd_rpc_sym.c
 		fi
